requirements.txt
httplib2==0.22.0
numpy==1.23.4
opencv-python==4.7.0.72
openvino==2022.3.0
requests==2.31.0
python-dotenv==1.0.0
deep-sort-realtime==1.3.2
pandas==2.0.2